tests/wpt/tests/import-maps/script-supports-importmap.html

<!DOCTYPE html>
<html>
<title>HTMLScriptElement.supports importmap</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_true(HTMLScriptElement.supports('importmap'));
}, 'HTMLScriptElement.supports returns true for \'importmap\'');

test(function() {
  assert_false(HTMLScriptElement.supports(' importmap'));
  assert_false(HTMLScriptElement.supports('importmap '));
  assert_false(HTMLScriptElement.supports('Importmap'));
  assert_false(HTMLScriptElement.supports('ImportMap'));
  assert_false(HTMLScriptElement.supports('importMap'));
  assert_false(HTMLScriptElement.supports('import-map'));
  assert_false(HTMLScriptElement.supports('importmaps'));
  assert_false(HTMLScriptElement.supports('import-maps'));
}, 'HTMLScriptElement.supports returns false for unsupported types');

</script>
